Rework exports of `typecheck` module.
