+++ local-modules/doc-store/BaseDocStore.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import Typecheck from 'typecheck';
+import { Typecheck } from 'typecheck';
 
 import BaseDoc from './BaseDoc';
