+++ local-modules/util-common/PromCondition.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import Typecheck from 'typecheck';
+import { Typecheck } from 'typecheck';
 
 /**
  * Promise that is permanently resolved as `true`. Used as a result from
